make it better

+++ db/migrate/20160110215052_add_column_to_users.rb
@@ -2,6 +2,6 @@ class AddColumnToUsers < ActiveRecord::Migration
   def change
     add_column :users, :total_amount, :float
     add_column :users, :current_amount, :float
-    add_column :users, :total_expenditures_amount, :double
+    add_column :users, :total_expenditures_amount, :float
   end
 end
